Fix panel resizing and map interaction issues.
