+++ client/package.json
@@ -16,7 +16,8 @@
     "react": "^19.0.0-rc.1",
     "react-dom": "^19.0.0-rc.1",
     "react-icons": "^5.4.0",
-    "react-leaflet": "^5.0.0-rc.2"
+    "react-leaflet": "^5.0.0-rc.2",
+    "recharts": "^3.1.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
