+++ client/src/components/bottom-panel/BottomPanel.css
@@ -1,19 +1,9 @@
-.bottom-panel-overlay {
-  position: fixed;
-  left: 30%;
-  right: 0;
-  bottom: 0;
-  top: 0;
-  pointer-events: auto;
-  z-index: 9999;
-}
-
 .bottom-panel {
   position: absolute;
   left: 0;
   right: 0;
   bottom: 0;
-  min-height: 60px;
+  min-height: 16px;
   max-height: 100%;
   background: #222;
   color: #fff;
@@ -27,9 +17,12 @@
 }
 
 .bottom-panel-drag {
-  height: 12px;
+  height: 16px;
   cursor: ns-resize;
   background: #444;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
 
 .bottom-panel-close {
