+++ client/src/components/map/Map.tsx
@@ -3,7 +3,7 @@ import { MapContainer, TileLayer, GeoJSON, ScaleControl } from 'react-leaflet'
 import { useQuery } from '@tanstack/react-query'
 import { useNavigate } from '@tanstack/react-router'
 import { MapEffect } from '../../utils/map/MapEffectUtil'
-import { WatershedIDContext } from '../../utils/watershedID/WatershedIDContext'
+import { WatershedIDContext } from '../../utils/watershed-id/WatershedIDContext'
 import { fetchChannels, fetchSubcatchments, fetchWatersheds } from '../../api/api'
 import WatershedToggle from './controls/WatershedToggle/WatershedToggle'
 import ZoomInControl from './controls/ZoomIn/ZoomIn'
