+++ client/src/components/side-panels/watershed/WatershedOverview.tsx
@@ -2,7 +2,7 @@ import { useContext, useMemo } from 'react';
 import { useNavigate } from '@tanstack/react-router'
 import { useQuery } from '@tanstack/react-query';
 import { fetchWatersheds } from '../../../api/api';
-import { WatershedIDContext } from '../../../utils/watershedID/WatershedIDContext';
+import { WatershedIDContext } from '../../../utils/watershed-id/WatershedIDContext';
 import AccordionItem from '../../accordian-item/AccordianItem'
 import './Watershed.css'
