+++ client/src/pages/home/Home.tsx
@@ -1,7 +1,8 @@
 import { Outlet } from '@tanstack/react-router';
 import { useContext } from 'react';
-import { WatershedIDContext } from '../../utils/watershedID/WatershedIDContext';
-import { BottomPanelProvider, useBottomPanelContext } from '../../components/bottom-panel/BottomPanelContext';
+import { WatershedIDContext } from '../../utils/watershed-id/WatershedIDContext';
+import { useBottomPanelContext } from '../../utils/bottom-panel/BottomPanelContext';
+import { BottomPanelProvider } from '../../utils/bottom-panel/BottomPanelProvider';
 import Map from '../../components/map/Map';
 import HomeSidePanelContent from '../../components/side-panels/home-info/HomeInfoPanel';
 import BottomPanel from '../../components/bottom-panel/BottomPanel';
