+++ client/src/routes/router.tsx
@@ -1,5 +1,5 @@
 import { lazyRouteComponent, createRoute, createRootRoute, createRouter, Outlet } from '@tanstack/react-router';
-import { WatershedIDProvider } from '../utils/watershedID/WatershedIDProvider';
+import { WatershedIDProvider } from '../utils/watershed-id/WatershedIDProvider';
 import Navbar from '../components/navbar/Navbar';
 import Home from '../pages/home/Home';
 import WatershedOverview from '../components/side-panels/watershed/WatershedOverview';
